IntroductionToDatabases/bookstore_project/backend/requirements.txt
Flask==2.0.1
Flask-Cors==3.0.10
psycopg2-binary==2.9.5
Flask-SQLAlchemy==2.5.1
python-dotenv==1.0.0
Werkzeug==2.0.3
MarkupSafe==2.0.1
Jinja2==3.0.3
itsdangerous==2.0.1
SQLAlchemy<2.0